requirements.txt
nvidia-cosmos-sdk
torch
transformers